requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
scipy==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
email-validator==2.2.0
loguru==0.7.2
pytest==8.0.2
httpx==0.27.0
openpyxl==3.1.2
python-dotenv==1.0.1
